requirements.txt
python-telegram-bot==20.7
python-dotenv==1.0.0
aiohttp==3.9.1
requests==2.31.0
gunicorn==21.2.0
openai==1.3.0
flask==3.0.0
werkzeug==3.0.1
fpdf2==2.7.9
pyswisseph==2.10.3.2
geopy==2.4.1
reportlab==4.4.4
pytz==2024.1
timezonefinder==6.2.0
Pillow==10.1.0
psycopg2-binary==2.9.9
